tests/wpt/tests/selection/modify-extend-word-trailing-inline-block.tentative.html

<!doctype html>
<meta charset=utf-8>
<title>Selection.modify: Extending a word towards the end of a line ended by an inline-block.</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<link rel="author" href="mailto:emilio@crisal.io" title="Emilio Cobos Ãlvarez">
<link rel="author" href="https://mozilla.org" title="Mozilla">
<link rel="help" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1698705">

<div id="block">This is a line<div style="display:inline-block"></div></div>

<script>
test(function() {
  const selection = getSelection();
  const block = document.getElementById("block");
  selection.collapse(block.childNodes[0], 0);
  assert_equals(selection.toString(), "", "Empty selection at beginning");
  for (let i = 0; i < 4; ++i) {
    selection.modify("extend", "forward", "word");
  }
  assert_equals(selection.toString(), "This is a line", "Should've found the four words");
});
</script>
